GameObject menus improvements
-you can now select gameobjects from the gameobjects tree window
-the selected gameobject gets displayed on the gameobject window
-you can now change the transform and activate or deactivate meshes from the gameobject window

+++ Based_Engine/Based_Engine/GameObject.h
@@ -13,7 +13,7 @@ struct Transform
 		z = 0;
 	}
 
-	int x, y, z;
+	float x, y, z;
 };
 
 class Component;
